fix: use bye for InternalIds

+++ NoxRelay/src/Instances/InstanceManager.cs
@@ -4,7 +4,7 @@ public class InstanceManager
 {
     public static List<Instance> Instances = new();
 
-    public static Instance Get(ushort internalId) => Instances.Find(instance => instance.InternalId == internalId);
+    public static Instance Get(byte internalId) => Instances.Find(instance => instance.InternalId == internalId);
     public static Instance Get(uint masterId) => Instances.Find(instance => instance.MasterId == masterId);
     public static void Add(Instance instance) => Instances.Add(instance);
     public static bool Has(ushort internalId) => Instances.Exists(instance => instance.InternalId == internalId);
@@ -15,9 +15,9 @@ public class InstanceManager
             Instances.Remove(instance);
     }
 
-    public static ushort GetNextInternalId()
+    public static byte GetNextInternalId()
     {
-        ushort internalId = 0;
+        byte internalId = 0;
         while (Has(internalId))
             internalId++;
         return internalId;
